Rename file to follow PEP8
- See https://peps.python.org/pep-0008/#package-and-module-names
