+++ .pre-commit-config.yaml
@@ -85,7 +85,7 @@ repos:
         language: system
         types_or: [python, jupyter]
         require_serial: true
-  - repo: https://github.com/pre-commit/mirrors-prettier
-    rev: v3.1.0
-    hooks:
-      - id: prettier
+#  - repo: https://github.com/pre-commit/mirrors-prettier # Does not work on Dapla and DaplaLab
+#    rev: v3.1.0
+#    hooks:
+#      - id: prettier
