+++ src/ssb_wrapr/__init__.py
@@ -7,12 +7,12 @@ from .lazy_rexpr import lazily
 from .library import importr
 from .library import library
 from .load_namespace import try_load_namespace
-from .RArray import RArray
-from .RDataFrame import RDataFrame
+from .rarray import RArray
+from .rdataframe import RDataFrame
 from .renv import Renv
-from .RFactor import RFactor
-from .RList import RDict
-from .RList import RList
-from .RView import RView
+from .rfactor import RFactor
+from .rlist import RDict
+from .rlist import RList
+from .rview import RView
 from .settings import Settings
 from .settings import settings
