+++ src/ssb_wrapr/renv.py
@@ -10,7 +10,7 @@ from .function_wrapper import rfunc  # wrap_rfunc should perhaps be its own modu
 from .function_wrapper import wrap_rfunc  # wrap_rfunc should perhaps be its own module
 from .load_namespace import try_load_namespace
 from .rutils import rcall
-from .RView import RView
+from .rview import RView
 from .settings import settings
 from .utils import ROutputCapture
 from .utils import pinfo
