+++ src/ssb_wrapr/rfactor.py
@@ -5,7 +5,7 @@ import rpy2.robjects as ro
 import rpy2.robjects.vectors as vc
 from rpy2.robjects import pandas2ri
 
-from .RAttributes import get_Rattributes
+from .rattributes import get_Rattributes
 
 
 class RFactor(pd.Series):
