+++ tests/test_gausssuppressionfromdata.py
