+++ tests/test_subsetting_rarray.py
